test 22

+++ 23213/v.abramenko/tests/lab22.c
@@ -3,6 +3,7 @@
 #include <fcntl.h>
 #include <unistd.h>
 #include <poll.h>
+#include <termios.h>
 
 #define TIMEOUT 5000
 
@@ -25,7 +26,8 @@ void add_file(cycle* cycle, int fd, char* filename) {
         exit(-1);
     }
     new_node->pfd.fd = fd;
-    new_node->pfd.events = POLLIN;
+    new_node->pfd.events = PU-OS-2024/23213/v.abramenko/lab22$ ssh v.abramenko@ccfit.nsu.ru
+OLLIN;
     new_node->filename = filename;
     new_node->next = new_node;
     new_node->prev = new_node;
@@ -47,6 +49,10 @@ void next(cycle* cycle) {
 }
 
 void remove_file(cycle* cycle) {
+    if (cycle->current == NULL) {
+        return;
+    }
+    close(cycle->current->pfd.fd);
     if (cycle->current->next == cycle->current) {
         cycle->head = NULL;
         free(cycle->current);
@@ -63,8 +69,34 @@ void remove_file(cycle* cycle) {
     }
 }
 
+void set_new_attr(int fd, struct termios* original_tio) {
+    if (isatty(fd) == 0) {
+        return;
+    }
+
+    struct termios new_tio;
+    new_tio = *original_tio;
+    new_tio.c_lflag |= ICANON;
+    new_tio.c_lflag |= ECHO;
+    if (tcsetattr(fd, TCSANOW, &new_tio) == -1) {
+       perror("tcsetattr failed");
+       exit(-1);
+    }
+}
+
+void set_old_attr(int fd, struct termios* original_tio) {
+    if (isatty(fd) == 0) {
+        return;
+    }
+
+    if (tcsetattr(fd, TCSANOW, original_tio) == -1) {
+       perror("tcsetattr failed");
+       exit(-1);
+    }
+}
+
 int main(int argc, char** argv) {
-    if (argc < 2){
+    if (argc < 2) {
         fprintf(stderr, "not enough arguments\n");
         exit(-1);
     }
@@ -73,8 +105,7 @@ int main(int argc, char** argv) {
     cycle.current = NULL;
     cycle.head = NULL;
 
-    for (int i = 1; i < argc; i++)
-    {
+    for (int i = 1; i < argc; i++) {
         int fd = open(argv[i], O_RDONLY);
         if (fd == -1) {
             perror("open failed");
@@ -88,27 +119,48 @@ int main(int argc, char** argv) {
         if (cycle.head == NULL) {
             break;
         }
-        
         printf("[%s]\n", cycle.current->filename);
+        
+        struct termios original_tio;
+        if (isatty(cycle.current->pfd.fd) != 0) {
+            if (tcgetattr(cycle.current->pfd.fd, &original_tio) == -1) {
+               perror("tcgetattr failed");
+               exit(-1);
+            }
+            set_new_attr(cycle.current->pfd.fd, &original_tio);
+        }
+
         int ret = poll(&cycle.current->pfd, 1, TIMEOUT);
         switch (ret) {
             case -1:
+                set_old_attr(cycle.current->pfd.fd, &original_tio);
                 perror("poll failed");
                 exit(-1);
             case 0:
                 printf("[%s] timeout\n\n", cycle.current->filename);
+                set_old_attr(cycle.current->pfd.fd, &original_tio);
                 break;
             default:
                 if (cycle.current->pfd.revents & POLLIN) {
-                    int bytes_read = read(cycle.current->pfd.fd, buf, sizeof(buf) - 1);
+                    ssize_t bytes_read = read(cycle.current->pfd.fd, buf, sizeof(buf) - 1);
                     if (bytes_read > 0) {
                         buf[bytes_read] = 0;
                         printf("from [%s] readed: %s\n", cycle.current->filename, buf);
+                        set_old_attr(cycle.current->pfd.fd, &original_tio);
                     } else {
-                        printf("[%s] closed\n\n", cycle.current->filename);
-                        close(cycle.current->pfd.fd);
+                        if (bytes_read != 0) {
+                            set_old_attr(cycle.current->pfd.fd, &original_tio);
+                            perror("read failed");
+                            exit(-1);
+                        }
+                        set_old_attr(cycle.current->pfd.fd, &original_tio);
+                        printf("end of file: [%s] closed\n\n", cycle.current->filename);
                         remove_file(&cycle);
                     }
+                } else {
+                    set_old_attr(cycle.current->pfd.fd, &original_tio);
+                    printf("[%s] closed\n\n", cycle.current->filename);
+                    remove_file(&cycle);
                 }
                 break;
         }
